programs_ntt/requirements.txt
george
sncosmo
dask
lightgbm
optuna
mlflow
keras
astropy
cesium
xgboost
tfmpl
umap-learn
bhtsne
dm-sonnet
bson
hyperopt
seaborn
matplotlib
tqdm
tensorflow-gpu
luigi
xarray
numpy
scikit-learn
pandas
click
netCDF4
bottleneck